python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "braintrust-otel-utils"
version = "0.1.0"
description = "OpenTelemetry utilities for filtering spans while preserving LLM-related spans"
authors = [{name = "Braintrust"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
py-modules = ["span_processor"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"